Added IOInterface and Distribution phase
Created a IOInterface class that simulates reading sequentialy from the disk with buffers for disk pages to read and write. It also contains a counter for disk accesses. The distribution algorithm is the setup for the polyphase merge sorting algorithm. Also added sample data and distribution results from the data.